+++ src/components/SelectGroupBy/SelectGroupBy.js
@@ -1,5 +1,4 @@
 import React, { Component } from 'react';
-import './SelectGroupBy.css';
 
 class SelectGroupBy extends Component {
 	constructor(props) {
