+++ src/modules/CPMUCalc.js
